refactor: reorganize project structure and update documentation links
