--- docs/Docker-GUI.md
@@ -1,1 +1,0 @@
-See [docker-gui-wiki-full.md](docker-gui-wiki-full.md)
