--- docs/Docker.md
@@ -1,1 +1,0 @@
-See [docker-wiki-full.md](docker-wiki-full.md)
