--- docs/Home.md
@@ -1,10 +1,0 @@
-[Configuration](Configuration.md)
-
-[Docker](Docker.md)
-
-[Install tips for Unraid](Install-and-tips-for-Unraid.md)
-
-[Seedbox / Linux Install](seedbox.md)
-[Windows Install](windows-install.md)
-
-[ffmpeg / max worker issues](ffmpeg---max-workers-issues.md)
