--- docs/Install-and-tips-for-Unraid.md
@@ -1,1 +1,0 @@
-See [unraid-wiki-full.md](unraid-wiki-full.md)
