+++ docs/assets/logo.png
