+++ docs/home.md
@@ -1,0 +1,10 @@
+[Configuration](configuration.md)
+
+[Docker](docker.md)
+
+[Install tips for Unraid](unraid.md)
+
+[Seedbox / Linux Install](seedbox.md)
+[Windows Install](windows-install.md)
+
+[ffmpeg / max worker issues](ffmpeg-max-workers-issues.md)
