+++ pyproject.toml
@@ -51,7 +51,7 @@ ignore = [
 "src/trackers/UNIT3D/UNIT3D_TEMPLATE.py" = ["N999"]
 
 [tool.ruff.lint.isort]
-known-first-party = ["cogs", "data", "src", "web_ui"]
+known-first-party = ["data", "src", "web_ui"]
 section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
 
 [tool.pyright]
