+++ src/trackers/UNIT3D/hawkeuno.py
@@ -6,7 +6,7 @@ import aiofiles
 import httpx
 from rich.markup import escape
 
-from cogs.redaction import Redaction
+from src.cogs.redaction import Redaction
 from src.console import logger
 from src.get_desc import DescriptionBuilder
 from src.languages import languages_manager
