+++ src/trackers/makingoff.py
@@ -18,7 +18,7 @@ import langcodes
 import pycountry
 from bs4 import BeautifulSoup
 
-from cogs.redaction import Redaction
+from src.cogs.redaction import Redaction
 from src.config_helpers import format_terminal_link
 from src.console import logger, prompt_in_thread
 from src.cookie_auth import CookieValidator
