+++ src/trackers/morethantv.py
@@ -14,7 +14,7 @@ import pyotp
 from defusedxml import ElementTree
 from rich.markup import escape
 
-from cogs.redaction import Redaction
+from src.cogs.redaction import Redaction
 from src.console import console, logger
 from src.get_desc import DescriptionBuilder
 from src.meta import Meta
