+++ src/trackers/totheglory.py
@@ -11,7 +11,7 @@ import httpx
 from bs4 import BeautifulSoup
 from unidecode import unidecode
 
-from cogs.redaction import Redaction
+from src.cogs.redaction import Redaction
 from src.console import console, logger
 from src.cookie_auth import CookieValidator
 from src.exceptions import *  # noqa #F405
